+++ crates/miden-objects/src/testing/note.rs
@@ -14,7 +14,7 @@ use crate::{
     account::AccountId,
     asset::Asset,
     note::{
-        Note, NoteAssets, NoteExecutionHint, NoteInputs, NoteMetadata, NoteRecipient, NoteScript,
+        Note, NoteAssets, NoteExecutionHint, NoteMetadata, NotePayload, NoteRecipient, NoteScript,
         NoteTag, NoteType,
     },
 };
@@ -63,11 +63,11 @@ impl NoteBuilder {
     /// Set the note's input to `inputs`.
     ///
     /// Note: This overwrite the inputs, the previous input values are discarded.
-    pub fn note_inputs(
+    pub fn note_payload(
         mut self,
         inputs: impl IntoIterator<Item = Felt>,
     ) -> Result<Self, NoteError> {
-        let validate = NoteInputs::new(inputs.into_iter().collect())?;
+        let validate = NotePayload::new(inputs.into_iter().collect())?;
         self.inputs = validate.into();
         Ok(self)
     }
@@ -132,7 +132,7 @@ impl NoteBuilder {
             self.note_execution_hint,
             self.aux,
         )?;
-        let inputs = NoteInputs::new(self.inputs)?;
+        let inputs = NotePayload::new(self.inputs)?;
         let recipient = NoteRecipient::new(self.serial_num, note_script, inputs);
 
         Ok(Note::new(vault, metadata, recipient))